refactor(core): group glyph post-processing and pipeline stages
Move validate/preprocess, flattened-SVG parsing, same-color merging and
placement into src/core/glyph. Add pipeline/prepare.ts (prepareSvgLayers:
validate → preprocess → flatten → parse → evenodd → merge) and rename
the orchestrator to runFontPipeline, keeping runPipeline exported as an
alias. Delete picoFromFile, writeLayerSvg and parseCodepointFromFilename.
Drop the .js import suffixes (the package compiles to CommonJS, where
they were never needed) and the jest mapper that stripped them. Update
docs/PIPELINE.md to the new layout.
