+++ packages/react-native-nano-icons/__tests__/helpers/golden.ts
@@ -5,7 +5,7 @@ import fsp from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
 
-import { runPipeline } from '../../src/core/pipeline/run';
+import { runFontPipeline } from '../../src/core/pipeline/index';
 import type { NanoGlyphMap, NanoLogger } from '../../src/core/types';
 
 const TEST_ICONS = path.resolve(__dirname, '../../test_icons');
@@ -95,7 +95,7 @@ export async function runOnIcons(opts: {
       await fsp.copyFile(abs, path.join(inputDir, file));
     }
 
-    await runPipeline(
+    await runFontPipeline(
       { ...PIPELINE, fontFamily: opts.fontFamily },
       { inputDir, outputDir, tempDir },
       { logger: quietLogger(warnings) }
