+++ packages/react-native-nano-icons/__tests__/pipeline.e2e.test.ts
@@ -5,7 +5,7 @@ import fsp from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
 
-import { runPipeline } from '../src/core/pipeline/run';
+import { runFontPipeline } from '../src/core/pipeline/index';
 import type { NanoGlyphMap } from '../src/core/types';
 
 // ---------------------------------------------------------------------------
@@ -35,7 +35,7 @@ describe('Pipeline E2E — outline (single-colour)', () => {
     outputDir = path.join(os.tmpdir(), `nano-e2e-${Date.now()}`);
     tempDir = path.join(os.tmpdir(), `nano-e2e-tmp-${Date.now()}`);
 
-    await runPipeline(
+    await runFontPipeline(
       {
         fontFamily: FONT_FAMILY,
         upm: UPM,
@@ -205,7 +205,7 @@ describe('Pipeline E2E — inputHash embedding', () => {
     outputDir = path.join(os.tmpdir(), `nano-e2e-hash-${Date.now()}`);
     tempDir = path.join(os.tmpdir(), `nano-e2e-hash-tmp-${Date.now()}`);
 
-    await runPipeline(
+    await runFontPipeline(
       {
         fontFamily: FONT_FAMILY,
         upm: UPM,
