+++ packages/react-native-nano-icons/__tests__/placement.unit.test.ts
@@ -1,6 +1,9 @@
 /** @jest-environment node */
 
-import { computePlacement, transformPathForFont } from '../src/core/svg/layers';
+import {
+  computePlacement,
+  transformPathForFont,
+} from '../src/core/glyph/placement';
 import type { PathKitModule } from '../src/core/pathkit/types';
 import { loadPathKit, flattenContours, signedArea } from './helpers/geometry';
