+++ packages/react-native-nano-icons/cli/link.ts
@@ -2,8 +2,8 @@ import fs from 'node:fs';
 import path from 'node:path';
 import * as plist from 'plist';
 import type { PBXNativeTarget, XcodeProject } from 'xcode';
-import type { NanoLogger } from './logger.js';
-import type { BuiltFont } from './build.js';
+import type { NanoLogger } from './logger';
+import type { BuiltFont } from './build';
 
 const ANDROID_FONTS_DIR = 'android/app/src/main/assets/fonts';
 const IOS_NANOICONS_FONTS_DIR = 'nanoicons-fonts';
