+++ packages/react-native-nano-icons/src/core/glyph/parse.ts
@@ -1,9 +1,22 @@
 import { DOMParser, type Element } from '@xmldom/xmldom';
 import { parseColor } from '../../utils/parse';
+import {
+  SVG_MOVE_PREFIX,
+  SVG_NUMBER,
+  SVG_TRAILING_CLOSE,
+  WHITESPACE,
+} from '../../utils/svgPatterns';
+
+export type ParsedPath = {
+  d: string;
+  fill: string | null;
+  fillRule?: 'evenodd';
+  noMerge?: boolean;
+};
 
 export type ParsedFlatSvg = {
   viewBox: [number, number, number, number];
-  paths: Array<{ d: string; fill: string | null; fillRule?: 'evenodd' }>;
+  paths: ParsedPath[];
 };
 
 // if the fill is implicit, walk ancestors for the first explicit fill value
@@ -30,19 +43,19 @@ export function calculateOpColor(
 }
 
 /**
- * If a flattened path lost its initial moveto (e.g. picosvg dropped an empty
- * `Mx y z` subpath), prepend `M` using the path's last coordinate pair.
+ * If a flattened path lost its initial moveto (e.g. an empty `Mx y z`
+ * subpath was dropped), prepend `M` using the path's last coordinate pair.
  * For closed icon shapes the endpoint equals the start point.
  */
 export function sanitizePathData(d: string): { d: string; sanitized: boolean } {
   const trimmed = d.trim();
-  if (!trimmed || /^[Mm]/.test(trimmed)) {
+  if (!trimmed || SVG_MOVE_PREFIX.test(trimmed)) {
     return { d: trimmed, sanitized: false };
   }
 
   // Strip trailing close commands, then grab the last two numbers as x,y
-  const withoutClose = trimmed.replace(/[Zz]\s*$/, '');
-  const nums = withoutClose.match(/-?\d+(?:\.\d+)?/g);
+  const withoutClose = trimmed.replace(SVG_TRAILING_CLOSE, '');
+  const nums = withoutClose.match(SVG_NUMBER);
   if (!nums || nums.length < 2) {
     return { d: trimmed, sanitized: false };
   }
@@ -52,15 +65,13 @@ export function sanitizePathData(d: string): { d: string; sanitized: boolean } {
   return { d: `M${x},${y} ${trimmed}`, sanitized: true };
 }
 
-export const parsePath = (
-  p: Element
-): { d: string; fill: string | null; fillRule?: 'evenodd' } => {
+const parsePath = (p: Element): ParsedPath => {
   const d = p.getAttribute('d') ?? '';
 
   const op = p.getAttribute('opacity');
   const fillOp = p.getAttribute('fill-opacity');
   const fill = p.getAttribute('fill');
-  // picosvg may drop fill-rule but preserve clip-rule; treat either as evenodd
+  // flattening may drop fill-rule but preserve clip-rule; treat either as evenodd
   const fillRule =
     p.getAttribute('fill-rule') === 'evenodd' ||
     p.getAttribute('clip-rule') === 'evenodd'
@@ -94,7 +105,7 @@ export function parseFlattenedSvg(
 
   const viewBoxRaw = svgEl
     ?.getAttribute('viewBox')
-    ?.split(/\s+/)
+    ?.split(WHITESPACE)
     .map(Number) ?? [0, 0, 100, 100];
 
   const viewBox: [number, number, number, number] =
@@ -120,28 +131,4 @@ export function shouldSkipPath(d: string, fill: string | null): boolean {
   if (!d || d.trim() === '') return true;
   const f = (fill ?? '').trim().toLowerCase();
   return f === 'transparent' || f === 'none';
-}
-
-export type SvgValidation = { valid: true } | { valid: false; reason: string };
-
-export function validateSvg(content: string): SvgValidation {
-  if (/<mask[\s>]/i.test(content)) {
-    return { valid: false, reason: '<mask> is not supported yet' };
-  }
-  if (/<filter[\s>]/i.test(content)) {
-    return { valid: false, reason: '<filter> is not supported yet' };
-  }
-  if (/<image[\s>]/i.test(content)) {
-    return {
-      valid: false,
-      reason: 'embedded raster <image> is not supported yet',
-    };
-  }
-  return { valid: true };
-}
-
-// ensure the svg has a xmlns attribute
-export function preprocessSvg(content: string): string {
-  if (/xmlns\s*=/.test(content)) return content;
-  return content.replace(/<svg\b/, '<svg xmlns="http://www.w3.org/2000/svg"');
 }
