+++ packages/react-native-nano-icons/src/core/glyph/placement.ts
@@ -1,6 +1,3 @@
-import fsp from 'node:fs/promises';
-import path from 'node:path';
-
 import type { PathKitModule } from '../pathkit/types';
 
 function roundInt(n: number): number {
@@ -36,35 +33,6 @@ export function computePlacement(opts: {
   const yOff = (opts.upm - scaledH) / 2; // ~= totalPadding
 
   return { vx, vy, scale, xOff, yOff, adv };
-}
-
-export async function writeLayerSvg(opts: {
-  tempDir: string;
-  upm: number;
-  adv: number;
-  vx: number;
-  vy: number;
-  scale: number;
-  xOff: number;
-  yOff: number;
-  d: string;
-  codepoint: number;
-}): Promise<void> {
-  const hex = opts.codepoint.toString(16).padStart(4, '0');
-
-  const layerSvg = `
-    <svg viewBox="0 0 ${opts.adv} ${
-      opts.upm
-    }" xmlns="http://www.w3.org/2000/svg">
-      <rect width="${opts.adv}" height="${opts.upm}" fill="none" />
-      <g transform="translate(${opts.xOff}, ${opts.yOff}) scale(${
-        opts.scale
-      }) translate(${-opts.vx}, ${-opts.vy})">
-        <path d="${opts.d}" fill="black" />
-      </g>
-    </svg>`.trim();
-
-  await fsp.writeFile(path.join(opts.tempDir, `u${hex}.svg`), layerSvg, 'utf8');
 }
 
 /**
