+++ packages/react-native-nano-icons/src/core/pipeline/index.ts
@@ -1,3 +1,7 @@
-export { runPipeline, type PipelineResult } from './run.js';
-export type { PipelineConfig, PipelinePaths } from './config.js';
-export { ensureEmptyDir } from './config.js';
+export {
+  runFontPipeline,
+  runFontPipeline as runPipeline,
+  type PipelineResult,
+} from './runFontPipeline';
+export type { PipelineConfig, PipelinePaths } from './config';
+export { ensureEmptyDir } from './config';
