+++ packages/react-native-nano-icons/src/core/tsconfig.json
@@ -26,9 +26,9 @@
   },
   "include": [
     "./pathkit/**/*.ts",
-    "./flatten/**/*.ts",
     "./pipeline/**/*.ts",
     "./font/**/*.ts",
-    "./svg/**/*.ts"
+    "./flatten/**/*.ts",
+    "./glyph/**/*.ts"
   ]
 }
