+++ packages/react-native-nano-icons/src/utils/parse.ts
@@ -1,12 +1,5 @@
 import { CSS_NAMED_COLORS } from '../const/colors';
 
-// e.g. u0041.svg -> 65
-export function parseCodepointFromFilename(filename: string): number {
-  const m = /^u([0-9a-fA-F]+)\.svg$/.exec(filename);
-  if (!m) throw new Error(`Unexpected glyph filename: ${filename}`);
-  return parseInt(m[1]!, 16);
-}
-
 // e.g. #ff0000 -> [255, 0, 0, 1], rgba(255, 0, 0, 0.5) -> [255, 0, 0, 0.5], rgb(255, 0, 0) -> [255, 0, 0, 1], blue -> [0, 0, 255, 1] etc.
 export function parseColor(
   color: string
